Adversarial review round 1: #41 Medium/Low findings
Six findings from the round-1 review of the batch-1 landing, all in
injuries-medical and its spec. One is a real arithmetic defect; the rest
are an unexercised guard and four documentation errors.

Sign inversion in AgeRiskFor. The age term computed

    (long)perYear * (ageYears - pivotYears)

which runs the subtraction in int and widens only the product. At an
extreme pivot the subtraction overflows and the term changes sign — an
injury-risk contribution pointing the wrong way. Widened to

    (long)perYear * ((long)ageYears - pivotYears)

Measured, not reasoned: AgeRiskFor(26, pivotYears: int.MinValue,
perYear: 150, span: 1800) returns +1800 with the fix and -1800 without
it. Locked by a test that fails when the widening is reverted.

The parameterised overload's negative-dial guard had no isolating case
anywhere and deleted with the whole suite green — the recurring class
this review keeps finding. Now locked by
AgeRiskFor_NegativeDials_FailLoud_TTMDAGE008, mutation-verified
(guard deleted -> Expected: InvalidOperationException, But was: null),
with spec id T-MD-AGE-008 allocated in section 5.

Documentation defects, all in text that was wrong rather than merely
stale:

  * perYear's param doc said "per-mille-of-a-million" — 1000x off the
    per-million scale the code uses.
  * AdvanceMedicalDay's exception docs omitted the negative-ageYears
    source and undercounted the consuming-site guards.
  * AgeRiskSpan's catalogue claimed its +/-12-year plateau "puts no
    cliff inside the football range". It is reachable today, because
    #28's retirement is flag-only with no roster removal, so long
    careers converge onto it. Recorded as knowingly accepted pending
    #28's retiree removal, rather than restating the false claim. The
    underlying #28 dependency is deliberately not touched here.
  * Appendix B's mid-recovery example was off by one against section
    3.6's own chain (day 208 is 4, not 3), and a second example carried
    a risk below the player's own floor.
  * The "27 unique, contiguous" FR-MD count claims, falsified by
    FR-MD-025a, corrected at the three named sites.

Recorded, not fixed: section-9-approval-checklist's R-02 and section
9.3.1 evidence rows still say "27 FRs". Outside the finding's three
named locations — the grep-boundary pattern this repo has hit five
times — and folded into the round's close-out instead.

MedicalStep.cs 1.14 -> 1.15; InjuriesMedicalConstants.cs 1.7 -> 1.8;
MedicalStepTests.cs 1.10 -> 1.11; section-5 0.8 -> 0.9;
section-9-approval-checklist 0.6 -> 0.7; outline 0.3 -> 0.4;
appendices 0.15 -> 0.16. SPEC_INDEX annotated in place.

InjuriesMedical.Tests 74 -> 76, all passing, verified independently at
this tree. SeasonSave.Tests 402 passed / 3 known skips — the locked
season-injury bands (717-816/season, starters 2.08, reserves 1.12,
9.4% unavailability) did not move. Lint 0 ERROR.

The whole tree does not build at this commit: player-progression is
mid-revision under a concurrent pass and AbilityModelTests.cs fails
CS1739. Nothing in this commit depends on it, and the two suites above
were run and passed against these exact files.

+++ src/injuries-medical/MedicalStep.cs
@@ -1,6 +1,6 @@
 // File:     src/injuries-medical/MedicalStep.cs
 // Created:  2026-08-05
-// Modified: 2026-08-22 (ERR-041-021 — AR over the ERR-041-020 landing, H4 — v1.14)
+// Modified: 2026-08-23 (Group-B AR findings over the ERR-041-021 landing — v1.15)
 // Author:   —
 // Spec:     Injuries & Medical #41 §3.1–§3.4 + Appendices A/B (FR-MD-003..016, FR-MD-023),
 //           F1/F4/F6/F7; Code Standards #20
@@ -63,7 +63,12 @@ namespace TacticalDirector.InjuriesMedical
         /// <param name="worldDay">The world day being advanced.</param>
         /// <param name="worldSeed">The world seed the draw key is derived from — the career's seed, not a per-match one.</param>
         /// <param name="occurrenceEnabled">The KD-8 dial. Off reduces the step to recovery-only with no draw at all (FR-MD-027).</param>
-        /// <exception cref="ArgumentOutOfRangeException"><paramref name="state"/> carries an undefined <see cref="InjurySeverity"/> (F4).</exception>
+        /// <exception cref="ArgumentOutOfRangeException">
+        /// <paramref name="state"/> carries an undefined <see cref="InjurySeverity"/> (F4); or
+        /// <paramref name="ageYears"/> is negative (from <see cref="AgeRiskFor(int)"/>, reached only on
+        /// an occurrence-eligible day — a derived age is never below zero, so a negative one is corrupt
+        /// state, not a young player).
+        /// </exception>
         /// <exception cref="ArgumentException">
         /// <paramref name="state"/> is incoherent — recovery outstanding while healthy, none while
         /// injured, negative, or above the ceiling (F1); either <paramref name="medical"/> multiplier
@@ -71,13 +76,15 @@ namespace TacticalDirector.InjuriesMedical
         /// last-advanced day (F7) or is itself the never-advanced sentinel.
         /// </exception>
         /// <exception cref="InvalidOperationException">
-        /// A catalogue/config integrity failure, not a caller error — one of the four consuming-site
+        /// A catalogue/config integrity failure, not a caller error — one of the FIVE consuming-site
         /// guards fired: <see cref="InjuriesMedicalConstants.InjuryRiskMax"/> outside
         /// <c>(0, OCCURRENCE_DRAW_DENOM]</c> (from <see cref="DrawOccurrence"/>);
         /// <see cref="InjuriesMedicalConstants.RecoveryDaysPerTickBase"/> non-positive or
-        /// <see cref="InjuriesMedicalConstants.RecoveryMax"/> below 1 (the countdown site); or the
+        /// <see cref="InjuriesMedicalConstants.RecoveryMax"/> below 1 (the countdown site); the
         /// severity split negative / summing past the denominator (from
-        /// <see cref="ClassifySeverityFromDraw"/>).
+        /// <see cref="ClassifySeverityFromDraw"/>); or
+        /// <see cref="InjuriesMedicalConstants.AgeRiskPerYearFromPivot"/> /
+        /// <see cref="InjuriesMedicalConstants.AgeRiskSpan"/> negative (from <see cref="AgeRiskFor(int)"/>).
         /// </exception>
         public static void AdvanceMedicalDay(
             ref InjuryState state,
@@ -477,8 +484,13 @@ namespace TacticalDirector.InjuriesMedical
         /// class of claim that gets falsified on first run (the ERR-008-021/-022 chain, three times).
         /// </summary>
         /// <param name="ageYears">The player's current age in whole years.</param>
-        /// <param name="pivotYears">The age at which the term is zero.</param>
-        /// <param name="perYear">Per-mille-of-a-million risk per year away from the pivot.</param>
+        /// <param name="pivotYears">
+        /// The age at which the term is zero. Deliberately UNGUARDED here (agerisk-int-subtraction-and-
+        /// both-dials): unlike <paramref name="perYear"/>/<paramref name="span"/>, no value of this
+        /// dial breaks a catalogue/config invariant — every pivot, including an extreme or mistyped one,
+        /// still produces a well-defined (if degenerate) term once the subtraction below is widened.
+        /// </param>
+        /// <param name="perYear">Risk contribution (per-million scale) per year away from the pivot.</param>
         /// <param name="span">Symmetric saturation magnitude.</param>
         internal static int AgeRiskFor(int ageYears, int pivotYears, int perYear, int span)
         {
@@ -496,7 +508,13 @@ namespace TacticalDirector.InjuriesMedical
                     + "inverts the clamp; catalogue/config integrity failure (§3.4, Appendix A).");
             }
 
-            long term = (long)perYear * (ageYears - pivotYears);
+            // Widened on BOTH operands (the ClassifySeverityFromDraw widening-comment standard,
+            // agerisk-int-subtraction-and-both-dials): with only the product widened, a sufficiently
+            // negative pivotYears overflows the int subtraction below before the cast ever runs — e.g.
+            // ageYears=26, pivotYears=int.MinValue computes 26 - int.MinValue in int arithmetic, which
+            // wraps to a large NEGATIVE value and inverts the term's sign league-wide. Widening the
+            // subtraction itself removes the wrap for any int-range pivot.
+            long term = (long)perYear * ((long)ageYears - pivotYears);
 
             return ClampLong(term, -span, span);
         }
@@ -739,4 +757,16 @@ namespace TacticalDirector.InjuriesMedical
 // |         |            |        | alike, and larger in relative terms for the more robust player). The
 // |         |            |        | normative position is restated as what is actually load-bearing: inside
 // |         |            |        | the sum, BEFORE the OccurrenceRiskMillMult scaling and BEFORE the clamp.
+// | 1.15    | 2026-08-23 | —      | Group-B AR findings (Medium guards-unexercised half + 3 Low). The
+// |         |            |        | parameterised AgeRiskFor's subtraction is widened on BOTH operands —
+// |         |            |        | (long)ageYears - pivotYears, not (ageYears - pivotYears) then cast — a
+// |         |            |        | sufficiently negative pivotYears otherwise overflows int and inverts the
+// |         |            |        | term's sign (agerisk-int-subtraction-and-both-dials); the pivotYears
+// |         |            |        | param doc now states it is deliberately the one unguarded dial and why.
+// |         |            |        | perYear's doc corrected from "per-mille-of-a-million" (1000x too small)
+// |         |            |        | to "per-million" (medicalstep-contract-docs). AdvanceMedicalDay's
+// |         |            |        | <exception> docs: the ArgumentOutOfRangeException tag now names the
+// |         |            |        | negative-ageYears source alongside the undefined-severity one, and the
+// |         |            |        | InvalidOperationException tag counts FIVE consuming-site guards (was
+// |         |            |        | four — AgeRiskFor's own guard is now named) rather than four.
 #endregion
